Inline Case files for LargeClassSpec and NestedBlockDepthSpec
