+++ detekt-rules-complexity/src/test/kotlin/io/gitlab/arturbosch/detekt/rules/complexity/NestedBlockDepthSpec.kt
@@ -1,6 +1,5 @@
 package io.gitlab.arturbosch.detekt.rules.complexity
 
-import io.github.detekt.test.utils.resourceAsPath
 import io.gitlab.arturbosch.detekt.api.ThresholdedCodeSmell
 import io.gitlab.arturbosch.detekt.test.TestConfig
 import io.gitlab.arturbosch.detekt.test.assertThat
@@ -17,7 +16,110 @@ class NestedBlockDepthSpec {
 
     @Test
     fun `should detect only the nested large class`() {
-        subject.lint(resourceAsPath("NestedClasses.kt"))
+        val code = """
+            class NestedClasses {
+            
+                private val i = 0
+            
+                class InnerClass {
+            
+                    class NestedInnerClass {
+            
+                        fun nestedLongMethod() {
+                            if (true) {
+                                if (true) {
+                                    if (true) {
+                                        5.run {
+                                            this.let {
+                                                listOf(1, 2, 3).map { it * 2 }
+                                                    .groupBy(Int::toString, Int::toString)
+                                            }
+                                        }
+                                    }
+                                }
+                            }
+            
+                            try {
+                                for (i in 1..5) {
+                                    when (i) {
+                                        1 -> print(1)
+                                    }
+                                }
+                            } finally {
+            
+                            }
+            
+                            fun nestedLocalMethod() {
+                                println()
+                                println()
+                                println()
+                                println()
+                                println()
+                                println()
+                                println()
+                                println()
+                                println()
+                                println()
+                                println()
+                                println()
+                                println()
+                                println()
+                                println()
+                                println()
+                                println()
+                                println()
+                                println()
+                                println()
+                                println()
+                                println()
+                                println()
+                                println()
+                                println()
+                                println()
+                                println()
+                                println()
+                                println()
+                                println()
+                                println()
+                                println()
+                                println()
+                                println()
+                                println()
+                                println()
+                                println()
+                                println()
+                                println()
+                                println()
+                                println()
+                                println()
+                                println()
+                                println()
+                                println()
+                                println()
+                                println()
+                                println()
+                                println()
+                                println()
+                                println()
+                                println()
+                                println()
+                                println()
+                                println()
+                                println()
+                            }
+                            nestedLocalMethod()
+                        }
+                    }
+                }
+            
+            }
+            
+            /**
+             * Top level members must be skipped for LargeClass rule
+             */
+            val aTopLevelPropertyOfNestedClasses = 0
+        """.trimIndent()
+        subject.lint(code)
         assertThat(subject.findings).hasSize(1)
         assertThat((subject.findings[0] as ThresholdedCodeSmell).value).isEqualTo(5)
     }
